Refactored instantiation
There is no point in having `new()` as part of a trait - unless the trait itself knows how to instantiate all variants of the class based on the params, but even then it should be a standalone function.

Instead, now both TCP and UDP have their own `new()` and `new_with_port(u16)`. Note that this is more in line with `mockito` design, except that there `with_port` has been deprecated in favor of `with_opts(ServerOpts)` -- which I think is too early. Eventually it may also need to be deprecated.

As part of the refactoring, renamed `listening_port` to `port`, and bumped minor version to indicate breaking API changes.  In a separate PR i will propose some CI changes to verify this automatically.